requirements.txt
flask
redis
flask_sqlalchemy
mysqlclient